fixed issue with users not being added / removed properly (#101)
This pull request includes changes to the `teamMemberList` and
`userList` components to enhance context menu functionality and improve
user management.

Enhancements to context menu functionality:

*
[`frontend/src/Components/teamMemberList.tsx`](diffhunk://#diff-106f46e9fe8397eaa6b06d3a0a42208e1a67d96165fafaa089b1df8a5ff5b6b5L60-R62):
Modified the `handleContextMenu` function to accept a username parameter
and updated the `onContextMenu` event to pass the username.
[[1]](diffhunk://#diff-106f46e9fe8397eaa6b06d3a0a42208e1a67d96165fafaa089b1df8a5ff5b6b5L60-R62)
[[2]](diffhunk://#diff-106f46e9fe8397eaa6b06d3a0a42208e1a67d96165fafaa089b1df8a5ff5b6b5L89-R89)
*
[`frontend/src/Components/userList.tsx`](diffhunk://#diff-af5ee1865ccd5d9b3db4da8223bef933460999f26012c2210941d812fc9d4b3eL57-R59):
Updated the `handleContextMenu` function to accept a username parameter
and adjusted the `onContextMenu` event to pass the username.
[[1]](diffhunk://#diff-af5ee1865ccd5d9b3db4da8223bef933460999f26012c2210941d812fc9d4b3eL57-R59)
[[2]](diffhunk://#diff-af5ee1865ccd5d9b3db4da8223bef933460999f26012c2210941d812fc9d4b3eL86-R85)

Improvements to user management:

*
[`frontend/src/Components/teamMemberList.tsx`](diffhunk://#diff-106f46e9fe8397eaa6b06d3a0a42208e1a67d96165fafaa089b1df8a5ff5b6b5L5-R5):
Added the `addUserToChannel` import to enable adding users to a channel.
*
[`frontend/src/Components/userList.tsx`](diffhunk://#diff-af5ee1865ccd5d9b3db4da8223bef933460999f26012c2210941d812fc9d4b3eL69):
Removed the commented-out menu item for setting a user to admin.