bug fix tasks #40 & #41: added userlist and teammemberlist refresh wh… (#103)
This pull request includes several changes to improve the user and team
management functionalities in both the backend and frontend of the
application. The most important changes include updating the
`listChannelUsers` method in the `DashboardService` class, adding a
refresh state to the `TeamMemberList` and `UserList` components, and
ensuring user lists are refreshed after certain actions.

Backend changes:

*
[`backend/src/services/dashboardService.ts`](diffhunk://#diff-d66ad31f67a614bc595c99565b502522c65bfd0b18cec392ffc5526981c96a23L64-R80):
Updated the `listChannelUsers` method to include user roles and sort the
users by their roles, with administrators appearing first.

Frontend changes:

*
[`frontend/src/Components/teamMemberList.tsx`](diffhunk://#diff-106f46e9fe8397eaa6b06d3a0a42208e1a67d96165fafaa089b1df8a5ff5b6b5R20-R23):
Added a `refreshState` prop to the `TeamMemberList` component to trigger
a refresh of the user list when necessary.
[[1]](diffhunk://#diff-106f46e9fe8397eaa6b06d3a0a42208e1a67d96165fafaa089b1df8a5ff5b6b5R20-R23)
[[2]](diffhunk://#diff-106f46e9fe8397eaa6b06d3a0a42208e1a67d96165fafaa089b1df8a5ff5b6b5L49-R50)
[[3]](diffhunk://#diff-106f46e9fe8397eaa6b06d3a0a42208e1a67d96165fafaa089b1df8a5ff5b6b5L74-R76)
*
[`frontend/src/Components/userList.tsx`](diffhunk://#diff-af5ee1865ccd5d9b3db4da8223bef933460999f26012c2210941d812fc9d4b3eR22-R25):
Added a `handleRefresh` prop to the `UserList` component to refresh the
user list after adding users to a team or channel.
[[1]](diffhunk://#diff-af5ee1865ccd5d9b3db4da8223bef933460999f26012c2210941d812fc9d4b3eR22-R25)
[[2]](diffhunk://#diff-af5ee1865ccd5d9b3db4da8223bef933460999f26012c2210941d812fc9d4b3eL67-R69)
*
[`frontend/src/Pages/AdminDashboard.tsx`](diffhunk://#diff-2bc287c9844e014ef9e4009f0514b00f57ef3166f8f3fe4870800b1fe373dfecR15):
Introduced a `refreshState` state and `handleRefresh` method to manage
the refresh state of the `TeamMemberList` and `UserList` components.
[[1]](diffhunk://#diff-2bc287c9844e014ef9e4009f0514b00f57ef3166f8f3fe4870800b1fe373dfecR15)
[[2]](diffhunk://#diff-2bc287c9844e014ef9e4009f0514b00f57ef3166f8f3fe4870800b1fe373dfecR24-R26)
[[3]](diffhunk://#diff-2bc287c9844e014ef9e4009f0514b00f57ef3166f8f3fe4870800b1fe373dfecR107)
[[4]](diffhunk://#diff-2bc287c9844e014ef9e4009f0514b00f57ef3166f8f3fe4870800b1fe373dfecR117)…en
user added/removed