fixed issue with admins not seeing the channels they are in, made adm… (#107)
This pull request includes several updates to the backend of the
application, focusing on enhancing user role validation and
authorization, as well as temporarily disabling rate limiting. The most
important changes include adding role-based authorization for removing
users from channels, updating the error handling in the
`ChannelController`, and modifying the `permissionMiddleware` to ensure
proper role validation.

### Role-based Authorization Enhancements:

*
[`backend/src/controllers/channelController.ts`](diffhunk://#diff-f1107bed0fa850f916d0136a1fc3e467b7605539430bed1506241efe73cab071L63-R63):
Updated the `removeUserFromChannel` method to include the `user.role`
parameter to ensure proper authorization when removing a user from a
channel.
*
[`backend/src/services/channelService.ts`](diffhunk://#diff-850d185c04ea70f3bd120a72c1bd5d14e4e50b58a551c8e4e4d7cc0a484bb85eL56-R56):
Added `userRole` parameter to `removeUserFromChannel` method and
included role-based checks to prevent non-SUPER_ADMIN users from
removing users if they are not an ADMIN.
[[1]](diffhunk://#diff-850d185c04ea70f3bd120a72c1bd5d14e4e50b58a551c8e4e4d7cc0a484bb85eL56-R56)
[[2]](diffhunk://#diff-850d185c04ea70f3bd120a72c1bd5d14e4e50b58a551c8e4e4d7cc0a484bb85eR67-R71)

### Error Handling Improvements:

*
[`backend/src/controllers/channelController.ts`](diffhunk://#diff-f1107bed0fa850f916d0136a1fc3e467b7605539430bed1506241efe73cab071R77-R78):
Added a new error handling condition to return a 403 status code when a
user is not authorized to remove another user from a channel.

### Middleware Updates:

*
[`backend/src/middlewares/permissionMiddleware.ts`](diffhunk://#diff-d2f371f6ff9210a586b3ae6328ee0cfdfdfedffcf6ab633dcd217eec6380956cL40-R41):
Adjusted the `checkTeamPermission` function to ensure the `teamMember`
is assigned before role validation.

### Temporary Rate Limiting Adjustment:

*
[`backend/src/server.ts`](diffhunk://#diff-03830d1462a5acd300db5f366d82d5232c32ff001e382f55d3cca74a65be60aaL37-R37):
Commented out the rate limiting middleware to potentially address
performance issues during testing.…ins not removable from a team/channel
unless by a super admin