Revert "Create README.md (#129)" (#131)
This pull request includes minor changes to the `README.md` files in
both `Sprint 1` and `Sprint 2` directories. The changes involve the
removal of an echo command that writes a message to `folder/README.md`.

Changes:

* [`Sprint
1/README.md`](diffhunk://#diff-cbe113f34297896a57f3b66103e86c7689415c7828bc55cebf3d3511248e4f4fL1):
Removed the echo command that writes "# This is the main folder" to
`folder/README.md`.
* [`Sprint
2/README.md`](diffhunk://#diff-cbe113f34297896a57f3b66103e86c7689415c7828bc55cebf3d3511248e4f4fL1):
Removed the echo command that writes "# This is the main folder" to
`folder/README.md`.This reverts commit
bfbed26dbb527167be4ecd9904ac62ea47994b59.