Channel enhancement (#192)
This pull request introduces several new features and improvements to
the ChatHavenApp. The most significant changes include adding the
ability for users to leave channels, defining default channels, and
enhancing the frontend to support these changes.

### Backend Changes:
* **ChannelController**: Added a new static method `leaveChannel` to
handle user requests to leave a channel. This method includes error
handling for various scenarios such as the channel or team not being
found, or the user not being a member of the team.
* **ChannelService**: Implemented the `leaveChannel` method to remove a
user from a channel and update the channel and team member records
accordingly.
* **Enums**: Introduced a new enumeration `DefaultChannels` to define
default channels such as 'General' and 'Random'.
* **SuperAdminService**: Modified the `createTeam` and `addUserToTeam`
methods to automatically create and add users to the default channels.
[[1]](diffhunk://#diff-78f2b23c3001ce0ccb4d63a9347f12a8d6e2c77646d8b93843c77d2bbd9c9eacR25-R28)
[[2]](diffhunk://#diff-78f2b23c3001ce0ccb4d63a9347f12a8d6e2c77646d8b93843c77d2bbd9c9eacR61-R64)

### Frontend Changes:
* **ChannelList Component**: Added context menu support to allow users
to leave channels. This includes new state management for context menu
visibility and position, and a new method `handleLeaveChannel` to call
the leave channel service. (F524c420L1,
[[1]](diffhunk://#diff-ea70798588ecf6ab4d587cfc9f76382c346e5e052ce0393aed63b4f55ee74b02L26-R35)
[[2]](diffhunk://#diff-ea70798588ecf6ab4d587cfc9f76382c346e5e052ce0393aed63b4f55ee74b02R88-R121)
[[3]](diffhunk://#diff-ea70798588ecf6ab4d587cfc9f76382c346e5e052ce0393aed63b4f55ee74b02L110-R173)
[[4]](diffhunk://#diff-ea70798588ecf6ab4d587cfc9f76382c346e5e052ce0393aed63b4f55ee74b02R186-R230)
* **ChannelService**: Added a new `leaveChannel` method to send a
request to the backend to leave a channel.

These changes enhance the user experience by providing more control over
channel membership and streamline the process of managing default
channels for teams.

---------

Co-authored-by: Shamsam <60717097+Shamsam@users.noreply.github.com>