Add space after username in MessageItem component (#228)
This pull request includes a small change to the
`ChatHavenApp/frontend/src/Components/Messaging/components/MessageItem.tsx`
file. The change adds a non-breaking space (`&nbsp;`) after the username
to improve the formatting of messages.

*
[`ChatHavenApp/frontend/src/Components/Messaging/components/MessageItem.tsx`](diffhunk://#diff-bd9efce0e9b3f640d429451e290f19733a198bfa2995b87bdfe773b2194d80e1L86-R86):
Added a non-breaking space after the `message.username` to ensure proper
spacing in the message display.

---------

Co-authored-by: M-ku5 <126633279+M-ku5@users.noreply.github.com>